Mainㅇ

+++ src/main/java/com/MGR/service/TicketService.java
@@ -1,14 +1,13 @@
 package com.MGR.service;
 
+import com.MGR.constant.ReservationStatus;
 import com.MGR.dto.ImageDto;
 import com.MGR.dto.MainTicketDto;
 import com.MGR.dto.TicketFormDto;
 import com.MGR.dto.TicketSearchDto;
-import com.MGR.entity.Image;
-import com.MGR.entity.Ticket;
+import com.MGR.entity.*;
 import com.MGR.exception.DuplicateTicketNameException;
-import com.MGR.repository.ImageRepository;
-import com.MGR.repository.TicketRepository;
+import com.MGR.repository.*;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -30,7 +29,9 @@ public class TicketService {
     private final TicketRepository ticketRepository;
     private final ImageRepository imageRepository;
     private final ImageService imageService;
-
+    private  final InventoryRepository inventoryRepository;
+    private final ReservationTicketRepository reservationTicketRepository;
+    private final OrderTicketRepository orderTicketRepository;
 @Transactional
     @Scheduled(cron = "0 0 0 * * ?") // 매일 자정에 실행
     public void deleteExpiredCoupons() {
@@ -45,7 +46,7 @@ public class TicketService {
     }
     
     public Long saveTicket(TicketFormDto ticketFormDto, List<MultipartFile> ticketImgFileList) throws Exception {
-        boolean isDuplicate = isDuplicateTicket(ticketFormDto);
+        boolean isDuplicate =isTicketNameDuplicated (ticketFormDto.getName());
         if (isDuplicate) {
             throw new DuplicateTicketNameException("중복된 티켓 정보가 존재합니다.");
         }
@@ -71,16 +72,11 @@ public class TicketService {
     }
 
 
-    private boolean isDuplicateTicket(TicketFormDto ticketFormDto) {
-        // TicketFormDto를 Ticket 엔티티로 변환
-        Ticket ticket = ticketFormDto.createTicket();
-
-        // TicketRepository를 사용하여 중복 여부 확인
-        Optional<Ticket> ticketOptional = ticketRepository.findByNameAndAdultPriceAndChildPriceAndMemoAndStartDateAndEndDateAndLocationCategory(
-                ticket.getName(), ticket.getAdultPrice(), ticket.getChildPrice(), ticket.getMemo(),
-                ticket.getStartDate(), ticket.getEndDate(), ticket.getLocationCategory()
-        );
-        return ticketOptional.isPresent();
+    public boolean isTicketNameDuplicated(String name) {
+        return ticketRepository.existsByName(name);
+    }
+    public boolean isExistedTicketNameDuplicated(String name, Long excludeId) {
+        return ticketRepository.existsByNameAndIdNot(name, excludeId);
     }
     //티켓 데이터를 읽어오는 함수
     @Transactional(readOnly = true)
@@ -99,12 +95,13 @@ public class TicketService {
         return ticketFormDto;
     }
 
+    @Transactional
 
     public Long updateTicket(TicketFormDto ticketFormDto, List<MultipartFile> ticketImgFileList) throws Exception {
-        boolean isDuplicate = isDuplicateTicket(ticketFormDto);
-        if (isDuplicate) {
-            throw new DuplicateTicketNameException("중복된 티켓 정보가 존재합니다.");
-        }
+//        boolean isDuplicate = isDuplicateTicket(ticketFormDto);
+//        if (isDuplicate) {
+//            throw new DuplicateTicketNameException("중복된 티켓 정보가 존재합니다.");
+//        }
 
         // 업데이트할 티켓을 가져옵니다.
         Ticket ticketToUpdate = ticketRepository.findById(ticketFormDto.getId())
@@ -125,17 +122,49 @@ public class TicketService {
         return ticketToUpdate.getId();
     }
 
+    @Transactional
+    public void deleteTicket(Long ticketId) {
+        try {
+            // ticketId로 모든 관련 예약 티켓을 가져옴
+            List<ReservationTicket> reservationTickets = reservationTicketRepository.findAllByTicketId(ticketId);
 
-    //삭제
-public void deleteTicket(Long ticketId) {
-    // 티켓에 연결된 이미지를 먼저 삭제
-    imageService.deleteImagesByTicketId(ticketId);
+            // reservationTickets가 비어 있거나, 모든 티켓의 상태가 RESERVE가 아닌 경우
+            boolean canDelete = reservationTickets.isEmpty() || reservationTickets.stream()
+                    .noneMatch(ticket -> ticket.getReservationStatus() == ReservationStatus.RESERVE);
 
-    // 티켓 삭제
-    ticketRepository.deleteById(ticketId);
-}
+            if (canDelete) {
+                // 티켓에 연결된 이미지 삭제
+                imageService.deleteImagesByTicketId(ticketId);
 
+                // 티켓과 관련된 모든 인벤토리 레코드 삭제
+                List<Inventory> inventories = inventoryRepository.findAllByTicketId(ticketId);
+                if (!inventories.isEmpty()) {
+                    inventoryRepository.deleteAll(inventories);
+                }
 
+                // ORDER_TICKET에서 티켓과 관련된 모든 레코드 삭제
+                List<OrderTicket> orderTickets = orderTicketRepository.findAllByTicketId(ticketId);
+                if (!orderTickets.isEmpty()) {
+                    orderTicketRepository.deleteAll(orderTickets);
+                }
+
+                // ReservationTicket 삭제
+                if (!reservationTickets.isEmpty()) {
+                    reservationTicketRepository.deleteAll(reservationTickets);
+                }
+
+                // 티켓 삭제
+                ticketRepository.deleteById(ticketId);
+            } else {
+                throw new IllegalStateException("상태가 'RESERVE'인 티켓이 있는 경우 티켓을 삭제할 수 없습니다.");
+            }
+        } catch (Exception e) {
+            // 예외 메시지 출력
+            System.out.println("Exception occurred: " + e.getMessage());
+            e.printStackTrace();
+            throw e;  // 예외를 다시 던져서 트랜잭션 롤백을 유발
+        }
+    }
 
     @Transactional(readOnly = true)
     public Page<Ticket> getAdminTicketPage(TicketSearchDto ticketSearchDto, Pageable pageable){
